code/mof-bot/requirements.txt
python-dotenv
numpy
rich
openai
pytest
